requirements.txt
pillow>=8.0.0
pyfiglet>=0.8.0
